아이템 드랍 및 사용 가능(버그있음)

+++ Assets/01.Scripts/100in/01_ItemLounge/WeaponStatusSO.cs
@@ -6,9 +6,9 @@ public class WeaponStatusSO : ScriptableObject
 {
     [SerializeField] private string _weaponName; public string WeaponName { get { return _weaponName; } }
     [SerializeField] private Sprite _weaponSprite; public Sprite WeaponSprite { get { return _weaponSprite; } }
-    [SerializeField] private int _keyValue; public int keyValue { get { return _keyValue; } }
-    [SerializeField] private int _hp; public int Hp { get { return _hp; } }
-    [SerializeField] private int _speed; public int Speed { get { return _speed; } }
-    [SerializeField] private int _attackPower; public int AttackPower { get { return _attackPower; } }
+    [SerializeField] private float _keyValue; public float keyValue { get { return _keyValue; } }
+    [SerializeField] private float _hp; public float Hp { get { return _hp; } }
+    [SerializeField] private float _speed; public float Speed { get { return _speed; } }
+    [SerializeField] private float _attackPower; public float AttackPower { get { return _attackPower; } }
     [SerializeField] private float _attackSpeed; public float AttackSpeed { get { return _attackSpeed; } }
 }
